EZP-24842: Refactored CleanInstaller to be MySQL-specific
